# mcp-sdk/spec/openapi-v1.yaml
openapi: 3.0.3
info:
  title: MCPLookup.org API
  version: 1.0.0
  description: |
    # MCPLookup.org API
    This enables intelligent server discovery and installation.

  contact:
    name: MCPLookup.org Support
    email: support@mcplookup.org
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: https://mcplookup.org/api/v1
    description: Production API v1
  - url: http://localhost:3000/api/v1
    description: Development API v1
security:
  - {}
  - BearerAuth: []
...
paths:
...
  /discover/smart:
    post:
      tags: [Discovery]
      summary: AI-powered smart discovery
      description: |
        Use AI to understand natural language queries and find the best matching servers.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                query:
                  type: string
                  description: Natural language query
                  example: "I need to work with files and folders in my project"
                context:
                  type: string
                  description: Additional context about the use case
                  example: "Building a Node.js application"
                max_results:
                  type: integer
                  default: 10
                  description: Maximum number of results
      responses:
        '200':
          description: Smart discovery results
          content:
            application/json:
              schema:
                type: object
                properties:
                  matches:
                    type: array
                    items:
                      type: object
                      properties:
                        server:
                          $ref: './schemas/mcp-server.yaml#/MCPServer'
                        relevance_score:
                          type: number
                          description: How well this server matches the query (0-1)
                        match_reasons:
                          type: array
                          items:
                            type: string
                          description: Why this server was recommended
                  query_analysis:
                    type: object
                    properties:
                      extracted_keywords:
                        type: array
                        items:
                          type: string
                      suggested_categories:
                        type: array
                        items:
                          type: string
                      intent:
                        type: string
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Optional API key for enhanced features
...